requirements.txt
# === Core Application Dependencies ===
numpy>=1.21.0
pandas>=1.3.0
jinja2>=3.0.0
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
torch>=1.9.0
transformers>=4.0.0

# === Evaluation & Metrics ===
deepeval>=0.1.0
ragas>=0.0.1
scikit-learn>=0.24.0
# Note: You also need to download the spaCy model separately
# Run: python -m spacy download en_core_web_sm

# === Dashboard & Visualization ===
streamlit
plotly>=5.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
reportlab

# === Development & Testing ===
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
pre-commit>=2.15.0